Control the number of client connecting to host to avoid timeout
