+++ SignalRServiceBenchmarkPlugin/plugins/Plugin.Microsoft.Azure.SignalR.Benchmark/SignalRPlugin/SlaveMethods/StartConnection.cs
@@ -23,7 +23,7 @@ namespace Plugin.Microsoft.Azure.SignalR.Benchmark.SlaveMethods
                 stepParameters.TryGetTypedValue(SignalRConstants.Type, out string type, Convert.ToString);
                 pluginParameters.TryGetTypedValue($"{SignalRConstants.ConnectionStore}.{type}", out IList<HubConnection> connections, (obj) => (IList<HubConnection>)obj);
 
-                await Task.WhenAll(BatchProcess(connections, StartConnect, concurrentConnection));
+                await Task.WhenAll(Util.BatchProcess(connections, StartConnect, concurrentConnection));
                 return null;
             }
             catch (Exception ex)
@@ -33,35 +33,7 @@ namespace Plugin.Microsoft.Azure.SignalR.Benchmark.SlaveMethods
                 throw;
             }
         }
-
-        private Task BatchProcess<T>(IList<T> source, Func<T, Task> f, int max)
-        {
-            var initial = (max >> 1);
-            var s = new System.Threading.SemaphoreSlim(initial, max);
-            _ = Task.Run(async () =>
-            {
-                for (int i = initial; i < max; i++)
-                {
-                    await Task.Delay(100);
-                    s.Release();
-                }
-            });
-
-            return Task.WhenAll(from item in source
-                                select Task.Run( async () =>
-                                {
-                                    await s.WaitAsync();
-                                    try
-                                    {
-                                        await f(item);
-                                    }
-                                    finally
-                                    {
-                                        s.Release();
-                                    }
-                                }));
-        }
-
+        
         private async Task StartConnect(HubConnection connection)
         {
             if (connection == null) return;
